disease-backend/requirements.txt
Flask==3.0.3
Flask_Cors==4.0.1
matplotlib==3.8.3
numpy==2.0.0
pandas==2.2.2
python-dotenv==1.0.1
scikit_learn==1.5.0
scipy==1.14.0
seaborn==0.13.2
python-dotenv